+++ demo/angular/src/app/gridstack.component.ts
@@ -1,5 +1,5 @@
 /**
- * gridstack.component.ts 7.1.2-dev
+ * gridstack.component.ts 7.2.0
  * Copyright (c) 2022 Alain Dumesny - see GridStack root license
  */